Se añadieron request response a cliente. Se realiza endpoint para buscar por tipo documento y numero.

+++ src/main/java/co/edu/usbcali/tiendaapp/repository/ClienteRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 @Repository
 public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
 
-    Optional<Cliente> findByTipoDocumentoIdAndDocumento(Long tipoDocumentoId, String documento);
+    Optional<Cliente> findByTipoDocumentoIdAndDocumento(Integer tipoDocumentoId, String documento);
 
 }
